Create userLikeOrder db

+++ server/models/applicantModel.js
@@ -34,7 +34,7 @@ const readApplicants = async orderId => {
         }
       ]
     })
-    .catch(err => err);
+    .catch(err => console.error(err));
 };
 
 const createOrderApply = async (orderId, userId, bidPrice, applyComment) => {
@@ -45,24 +45,26 @@ const createOrderApply = async (orderId, userId, bidPrice, applyComment) => {
       bidPrice,
       applyComment
     })
-    .catch(err => err);
+    .catch(err => console.error(err));
 };
 
 const readUserApplyOrNot = async (orderId, userId) => {
-  return await db.applicant.findOne({ where: { orderId, userId } }).catch(err => err);
+  return await db.applicant.findOne({ where: { orderId, userId } }).catch(err => console.error(err));
 };
 
 const updateOrderApply = async (orderId, userId, bidPrice, applyComment) => {
   console.log('bidPrice, applyComment', bidPrice, applyComment);
-  return await db.applicant.update({ bidPrice, applyComment }, { where: { orderId, userId } }).catch(err => err);
+  return await db.applicant
+    .update({ bidPrice, applyComment }, { where: { orderId, userId } })
+    .catch(err => console.error(err));
 };
 
 const deleteMyOrderApply = async (orderId, userId) => {
-  return await db.applicant.destroy({ where: { orderId, userId } }).catch(err => err);
+  return await db.applicant.destroy({ where: { orderId, userId } }).catch(err => console.error(err));
 };
 
 const deleteApplicants = async orderId => {
-  return await db.applicant.destroy({ where: { orderId } }).catch(err => err);
+  return await db.applicant.destroy({ where: { orderId } }).catch(err => console.error(err));
 };
 
 module.exports = {
